+++ src/services/telemetry/providers/posthog/PostHogClientProvider.ts
@@ -31,6 +31,7 @@ export class PostHogClientProvider {
 	/**
 	 * Filters PostHog events before they are sent.
 	 * For exceptions, we only capture those from the Cline extension.
+	 * this is specifically to avoid capturing errors from anything other than Cline
 	 */
 	static eventFilter(event: EventMessage | null) {
 		if (!event || event?.event !== "$exception") {
